refactor: 별도의 파일에 있던 이벤트핸들러 index.js로 이동
