--- src/pages/AddPages/AddFormPage/CardInfoForm/SecurityCodeInput/handler.js
@@ -1,11 +1,0 @@
-import { SECURITY_CODE_LENGTH } from '../../../../../constants';
-
-export const handleSecurityCodeInputChange = (props) => {
-  const { e, setSecurityCode, passwordInputRef } = props;
-  const slicedInputValue = e.target.value.slice(0, SECURITY_CODE_LENGTH);
-
-  if (slicedInputValue.length === SECURITY_CODE_LENGTH) {
-    passwordInputRef.current.focus();
-  }
-  setSecurityCode(slicedInputValue);
-};
